tests for specific node related things
