+++ src/zeyrho/btree/mod.rs
@@ -1,0 +1,6 @@
+pub mod tree;
+mod node;
+
+const DEGREE: usize = 3;
+const SEPARATORS_MAX_SIZE: usize = DEGREE - 1;
+const CHILDREN_MAX_SIZE: usize = DEGREE;
